remove common modules
