+++ app/build.gradle
@@ -67,7 +67,6 @@ android {
 }
 
 dependencies {
-    implementation project(':common-jvm')
 
     // Kotlin
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
