+++ app/src/main/java/io/github/droidkaigi/confsched2018/model/Topic.kt
