--- common-jvm/build.gradle
@@ -1,10 +1,0 @@
-apply plugin: 'kotlin-platform-jvm'
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:$kotlin_version"
-    expectedBy project(":common")
-}
-
-sourceCompatibility = "1.7"
-targetCompatibility = "1.7"
