--- common/build.gradle
@@ -1,7 +1,0 @@
-apply plugin: 'kotlin-platform-common'
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:$kotlin_version"
-}
-
