--- common/src/main/kotlin/io/github/droidkaigi/confsched2018/model/Date.kt
@@ -1,18 +1,0 @@
-package io.github.droidkaigi.confsched2018.model
-
-expect class Date() {
-    fun getDate(): Int
-    fun getMonth(): Int
-    fun getFullYear(): Int
-    fun getHours(): Int
-    fun getMinutes(): Int
-    fun getTime(): Number
-}
-
-expect operator fun Date.compareTo(otherDate: Date): Int
-
-expect fun parseDate(timeInMills: Long): Date
-expect fun Date.toReadableDateString(): String
-expect fun Date.toReadableTimeString(): String
-
-fun Date.toReadableDateTimeString() = "${toReadableDateString()} ${toReadableTimeString()}"
